--- app/views/index/home.scala.html
@@ -1,6 +1,0 @@
-@(name: models.CachedData)(implicit request: Request[_], flash: Flash, user: Option[CachedData], feedbackUrl: String, messages: Messages)
-
-@main_template(title = "Hello from fset-fasttrack-frontend") {
-
-    <h1>Hello @{name.user.firstName}</h1>
-}
