--- app/views/index/index.scala.html
@@ -1,11 +1,0 @@
-@(messages: Seq[String])(implicit request: Request[_], flash: Flash, user: Option[CachedData], feedbackUrl: String, msgs: Messages)
-
-@main_template(title = "Hello from fset-fasttrack-frontend") {
-    <h1>Hello from fset-fasttrack-frontend !</h1>
-    <section>Backends</section>
-    <ul>
-        @for(msg <- messages) {
-            <li>@msg</li>
-        }
-    </ul>
-}
