+++ conf/application.conf
@@ -18,7 +18,6 @@ include "silhouette.conf"
 
 appName="fset-fasttrack-frontend"
 play.modules.enabled += "com.kenshoo.play.metrics.PlayModule"
-play.modules.enabled += "modules.SilhouetteModule"
 play.http.requestHandler = "play.api.http.GlobalSettingsHttpRequestHandler"
 application.router=prod.Routes
 
@@ -31,6 +30,23 @@ controllers {
     needsAuditing = false
     needsLogging = false
   }
+}
+
+# Restore older play 2.3 CSRF behaviours
+play.filters.csrf {
+  header {
+    bypassHeaders {
+      X-Requested-With = "*"
+      Csrf-Token = "nocheck"
+    }
+    protectHeaders = null
+  }
+  bypassCorsTrustedOrigins = false
+  method {
+    whiteList = []
+    blackList = ["POST"]
+  }
+  contentType.blackList = ["application/x-www-form-urlencoded", "multipart/form-data", "text/plain"]
 }
 
 # IE8 will not render fonts when served over HTTPS unless they have a caching header,
